backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

langchain==0.3.18
langchain-community==0.3.17
langchain-openai==0.2.9

openai==1.59.3

pypdf==3.17.4
faiss-cpu>=1.8.0
python-dotenv==1.0.1
